change the structure, and update kfold modules
