+++ src/kfold/model/modules/__init__.py
@@ -1,8 +1,10 @@
-from . import (
-    confidence_head,
-    diffusion,
-    distogram_head,
-    seq_encoder,
-    struct_encoder,
+from . import (  # noqa
+    sequence_encoder,
+    structure_encoder,
+    input_embedder,
     trunk,
+    diffusion_module,
+    structure_module,
+    distogram_head,
+    confidence_head,
 )
