+++ src/kfold/model/modules/confidence_head/__init__.py
@@ -4,7 +4,6 @@ import importlib
 from pathlib import Path
 
 from . import *  # noqa
-from .base import BaseConfidenceHead
 
 # Get all Python module names in current directory
 _package_dir = Path(__file__).parent
