+++ src/kfold/model/modules/input_embedder/__init__.py
@@ -4,7 +4,7 @@ import importlib
 from pathlib import Path
 
 from . import *  # noqa
-from .base import BaseInputEmbedding
+from .base import BaseInputEmbedder
 
 # Get all Python module names in current directory
 _package_dir = Path(__file__).parent
