+++ src/kfold/model/modules/input_embedder/af3_embedder.py
@@ -4,11 +4,9 @@ import torch
 import torch.nn.functional as F
 
 import kfold.constants as C
-from kfold.boltz_local.modules.encoders import (
-    AtomAttentionEncoder,
-)
-from kfold.boltz_local.modules.utils import LinearNoBias
 from kfold.data.model_input import FoldingInput
+from kfold.model.layers.alphafold3.primitives import LinearNoBias
+from kfold.model.layers.alphafold3.transformers import AtomAttentionEncoder
 from kfold.utils.registry import INPUT_EMBEDDER, BaseConfig
 
 from .base import BaseInputEmbedder
@@ -55,16 +53,15 @@ class AF3InputEmbedder(BaseInputEmbedder):
             channel_token=cfg.channel_s,  # Same to channel_s
             atoms_per_window_queries=cfg.atoms_per_window_queries,
             atoms_per_window_keys=cfg.atoms_per_window_keys,
-            atom_encoder_depth=cfg.atom_encoder_depth,
-            atom_encoder_heads=cfg.atom_encoder_heads,
+            num_blocks=cfg.atom_encoder_depth,
+            num_heads=cfg.atom_encoder_heads,
         )
 
         # residue info
         self.num_res_types: int = C.NUM_RES_TYPES
-        self.num_pocket_contact_types: int = C.NUM_POCKET_CONTACT_TYPES
 
         # out projection
-        s_input_dim = cfg.channel_s + self.num_res_types + self.num_pocket_contact_types
+        s_input_dim = cfg.channel_s + self.num_res_types + 32 + 1
         self.s_init = LinearNoBias(s_input_dim, cfg.channel_s)
 
     def forward(self, f_input: FoldingInput) -> torch.Tensor:
@@ -80,23 +77,27 @@ class AF3InputEmbedder(BaseInputEmbedder):
         Tensor
             The embedded tokens. [L, c_s]
         """
+        # FIXME: add more
 
         # Atom attention encoder forward
         a, *_ = self.encoder(f_input)  # [L, c_s]
 
         # Concatenate additional token features
         res_type = f_input.token.res_type  # [L,]
-        pocket_constraint = f_input.token.pocket_contact_type  # [L,]
+        profile = f_input.msa.profile  # [L,]
+        deletion_mean = f_input.msa.deletion_mean  # [L,]
         s = torch.cat(
             [
                 a,
                 F.one_hot(res_type, self.num_res_types).float(),
-                F.one_hot(pocket_constraint, self.num_pocket_contact_types).float(),
+                F.one_hot(profile, 32).float(),
+                deletion_mean.unsqueeze(-1),
             ],
             dim=-1,
         )
 
         # Project to model dimension
+        # NOTE: (SeonghwanSeo) I introduce additional linear layer to unify the dimension.
         s = self.s_init(s)  # [L, c_s]
 
         return s
@@ -113,10 +114,10 @@ class AtomAttentionEncoderWithoutStructure(AtomAttentionEncoder):
         channel_atom: int,
         channel_atompair: int,
         channel_token: int,
+        num_blocks: int = 3,
+        num_heads: int = 4,
         atoms_per_window_queries: int = 32,
         atoms_per_window_keys: int = 128,
-        atom_encoder_depth=3,
-        atom_encoder_heads=4,
         activation_checkpointing=False,
     ):
         super().__init__(
@@ -125,10 +126,10 @@ class AtomAttentionEncoderWithoutStructure(AtomAttentionEncoder):
             channel_atom=channel_atom,
             channel_atompair=channel_atompair,
             channel_token=channel_token,
+            num_blocks=num_blocks,
+            num_heads=num_heads,
             atoms_per_window_queries=atoms_per_window_queries,
             atoms_per_window_keys=atoms_per_window_keys,
-            atom_encoder_depth=atom_encoder_depth,
-            atom_encoder_heads=atom_encoder_heads,
             use_structure=False,
             activation_checkpointing=activation_checkpointing,
         )
