+++ src/kfold/model/modules/trunk/pairformer.py
@@ -1,28 +1,16 @@
 import torch
-import torch.nn as nn
-from fairscale.nn.checkpoint.checkpoint_activations import checkpoint_wrapper
 
-from kfold.boltz_local.layers.attention import AttentionPairBias
-from kfold.boltz_local.layers.dropout import get_dropout_mask
-from kfold.boltz_local.layers.transition import Transition
-from kfold.boltz_local.layers.triangular_attention.attention import (
-    TriangleAttentionEndingNode,
-    TriangleAttentionStartingNode,
-)
-from kfold.boltz_local.layers.triangular_mult import (
-    TriangleMultiplicationIncoming,
-    TriangleMultiplicationOutgoing,
-)
+from kfold.model.layers.alphafold3.pairformer import PairformerStack
 from kfold.utils.registry import TRANSFORMER_MODULE
 
-from .base import BaseTransformerConfig
+from .base import BaseTransformer
 
 
 @TRANSFORMER_MODULE.register()
-class PairformerModule(nn.Module):
+class PairformerModule(BaseTransformer):
     """Pairformer module."""
 
-    class Config(BaseTransformerConfig):
+    class Config(BaseTransformer.Config):
         """Configuration for the Pairformer module.
 
         Parameters
@@ -58,50 +46,24 @@ class PairformerModule(nn.Module):
         pairwise_num_heads: int = 4
         activation_checkpointing: bool = False
         offload_to_cpu: bool = False
+        use_kernels: bool = False
 
-    def __init__(self, cfg: Config, use_kernel: bool = False):
+    def __init__(self, cfg: Config, use_kernels: bool = False):
         """Initialize the Pairformer module."""
-        super().__init__()
-        self.channel_s: int = cfg.channel_s
-        self.channel_z: int = cfg.channel_z
-        self.num_blocks: int = cfg.num_blocks
-        self.dropout: float = cfg.dropout
-        self.num_heads: int = cfg.num_heads
-        self.pairwise_head_width: int = cfg.pairwise_head_width
-        self.pairwise_num_heads: int = cfg.pairwise_num_heads
-        self.activation_checkpointing: bool = cfg.activation_checkpointing
-        self.offload_to_cpu: bool = cfg.offload_to_cpu
-        self.use_kernels: bool = use_kernel
+        super().__init__(cfg)
 
-        self.layers = nn.ModuleList()
-        for _ in range(cfg.num_blocks):
-            if cfg.activation_checkpointing:
-                self.layers.append(
-                    checkpoint_wrapper(
-                        PairformerLayer(
-                            self.channel_s,
-                            self.channel_z,
-                            self.num_heads,
-                            self.dropout,
-                            self.pairwise_head_width,
-                            self.pairwise_num_heads,
-                            use_kernels=self.use_kernels,
-                        ),
-                        offload_to_cpu=self.offload_to_cpu,
-                    )
-                )
-            else:
-                self.layers.append(
-                    PairformerLayer(
-                        self.channel_s,
-                        self.channel_z,
-                        self.num_heads,
-                        self.dropout,
-                        self.pairwise_head_width,
-                        self.pairwise_num_heads,
-                        use_kernels=self.use_kernels,
-                    )
-                )
+        self.trunk = PairformerStack(
+            channel_s=cfg.channel_s,
+            channel_z=cfg.channel_z,
+            num_blocks=cfg.num_blocks,
+            num_heads=cfg.num_heads,
+            dropout=cfg.dropout,
+            pairwise_head_width=cfg.pairwise_head_width,
+            pairwise_num_heads=cfg.pairwise_num_heads,
+            activation_checkpointing=cfg.activation_checkpointing,
+            offload_to_cpu=cfg.offload_to_cpu,
+            use_kernels=cfg.use_kernels or use_kernels,
+        )
 
     def forward(
         self,
@@ -116,21 +78,20 @@ class PairformerModule(nn.Module):
         Parameters
         ----------
         s : torch.Tensor
-            The sequence embeddings
+            Tensor of shape (B, L, c_s) containing token single feature
         z : torch.Tensor
-            The pairwise embeddings
+            Tensor of shape (B, L, L, c_z) containing token pair feature
         mask : torch.Tensor
-            The token mask
+            The token mask of shape (B, L)
         pair_mask : torch.Tensor
-            The pairwise mask
+            The pairwise mask of shape (B, L, L)
 
         Returns
         -------
-        torch.Tensor
-            The updated sequence embeddings.
-        torch.Tensor
-            The updated pairwise embeddings.
-
+        s_trunk: torch.Tensor
+            The updated tensor of shape (B, L, c_s).
+        z_trunk: torch.Tensor
+            The updated tensor of shape (B, L, L, c_z).
         """
         if not self.training:
             if z.shape[1] > 384:
@@ -140,116 +101,5 @@ class PairformerModule(nn.Module):
         else:
             chunk_size_tri_attn = None
 
-        for layer in self.layers:
-            s, z = layer(
-                s,
-                z,
-                mask,
-                pair_mask,
-                chunk_size_tri_attn,
-            )
-        return s, z
-
-
-class PairformerLayer(nn.Module):
-    """Pairformer module."""
-
-    def __init__(
-        self,
-        channel_s: int = 384,
-        channel_z: int = 128,
-        num_heads: int = 16,
-        dropout: float = 0.25,
-        pairwise_head_width: int = 32,
-        pairwise_num_heads: int = 4,
-        use_kernels: bool = False,
-    ):
-        """Initialize the Pairformer module.
-
-        Parameters
-        ----------
-        channel_s : int
-            The token single embedding size.
-        channel_z : int
-            The token pairwise embedding size.
-        num_heads : int, optional
-            The number of heads, by default 16
-        dropout : float, optional
-            The dropout rate, by default 0.25
-        pairwise_head_width : int, optional
-            The pairwise head width, by default 32
-        pairwise_num_heads : int, optional
-            The number of pairwise heads, by default 4
-        use_kernels : bool, optional
-            Whether to use custom kernels, by default False
-
-        """
-        super().__init__()
-        self.channel_s: int = channel_s
-        self.channel_z: int = channel_z
-        self.dropout: float = dropout
-        self.num_heads: int = num_heads
-        self.use_kernels: bool = use_kernels
-
-        self.tri_mul_out = TriangleMultiplicationOutgoing(channel_z)
-        self.tri_mul_in = TriangleMultiplicationIncoming(channel_z)
-        self.tri_att_start = TriangleAttentionStartingNode(
-            channel_z, pairwise_head_width, pairwise_num_heads, inf=1e9
-        )
-        self.tri_att_end = TriangleAttentionEndingNode(
-            channel_z, pairwise_head_width, pairwise_num_heads, inf=1e9
-        )
-
-        self.attention = AttentionPairBias(
-            channel_s, 0, channel_z, num_heads, use_s=False
-        )
-
-        self.transition_s = Transition(channel_s, channel_s * 4)
-        self.transition_z = Transition(channel_z, channel_z * 4)
-
-    def forward(
-        self,
-        s: torch.Tensor,
-        z: torch.Tensor,
-        mask: torch.Tensor,
-        pair_mask: torch.Tensor,
-        chunk_size_tri_attn: int | None = None,
-    ) -> tuple[torch.Tensor, torch.Tensor]:
-        """Perform the forward pass."""
-
-        # Line 2
-        dropout = get_dropout_mask(z, self.dropout, self.training)
-        z = z + dropout * self.tri_mul_out(z, mask=pair_mask)
-
-        # Line 3
-        dropout = get_dropout_mask(z, self.dropout, self.training)
-        z = z + dropout * self.tri_mul_in(z, mask=pair_mask)
-
-        # Line 4
-        dropout = get_dropout_mask(z, self.dropout, self.training)
-        z = z + dropout * self.tri_att_start(
-            z,
-            mask=pair_mask,
-            chunk_size=chunk_size_tri_attn,
-            use_kernels=self.use_kernels,
-        )
-
-        # Line 5
-        dropout = get_dropout_mask(z, self.dropout, self.training, columnwise=True)
-        z = z + dropout * self.tri_att_end(
-            z,
-            mask=pair_mask,
-            chunk_size=chunk_size_tri_attn,
-            use_kernels=self.use_kernels,
-        )
-
-        # Line 6
-        z = z + self.transition_z(z)
-
-        # Line 7
-        s = s + self.attention(s, None, z, mask)
-
-        # Line 8
-        s = s + self.transition_s(s)
-
-        return s, z
+        s_trunk, z_trunk = self.trunk(s, z, mask, pair_mask, chunk_size_tri_attn)
+        return s_trunk, z_trunk
